+++ packages/nhsuk-frontend/src/nhsuk/components/character-count/character-count.jsdom.test.mjs
@@ -1,9 +1,12 @@
 import { components } from '@nhsuk/frontend-lib'
 import { getByRole } from '@testing-library/dom'
 
-import { initCharacterCounts } from './character-count.mjs'
+import { CharacterCount, initCharacterCounts } from './character-count.mjs'
 
 describe('Character count', () => {
+  /** @type {HTMLElement} */
+  let $root
+
   /** @type {HTMLTextAreaElement} */
   let $textarea
 
@@ -21,16 +24,16 @@ describe('Character count', () => {
       }
     })
 
-    const $container = document.querySelector('.nhsuk-character-count')
+    $root = document.querySelector('.nhsuk-character-count')
 
-    $textarea = getByRole($container, 'textbox', {
+    $textarea = getByRole($root, 'textbox', {
       name: 'Can you provide more detail?'
     })
 
     jest.spyOn($textarea, 'addEventListener')
   })
 
-  describe('Initialisation', () => {
+  describe('Initialisation via init function', () => {
     it('should add event listeners', () => {
       initCharacterCounts()
 
@@ -63,6 +66,34 @@ describe('Character count', () => {
     it('should not throw with empty scope', () => {
       const scope = document.createElement('div')
       expect(() => initCharacterCounts({ scope })).not.toThrow()
+    })
+  })
+
+  describe('Initialisation via class', () => {
+    it('should not throw with $root element', () => {
+      expect(() => new CharacterCount($root)).not.toThrow()
+    })
+
+    it('should throw with unsupported browser', () => {
+      document.body.classList.remove('nhsuk-frontend-supported')
+
+      expect(() => new CharacterCount($root)).toThrow(
+        'NHS.UK frontend is not supported in this browser'
+      )
+    })
+
+    it('should throw with missing $root element', () => {
+      expect(() => new CharacterCount()).toThrow(
+        'CharacterCount: Root element (`$root`) not found'
+      )
+    })
+
+    it('should throw with wrong $root element type', () => {
+      $root = document.createElementNS('http://www.w3.org/2000/svg', 'svg')
+
+      expect(() => new CharacterCount($root)).toThrow(
+        'CharacterCount: Root element (`$root`) is not of type HTMLElement'
+      )
     })
   })
 })