+++ packages/nhsuk-frontend/src/nhsuk/components/checkboxes/checkboxes.jsdom.test.mjs
@@ -2,9 +2,12 @@ import { components } from '@nhsuk/frontend-lib'
 import { getByRole } from '@testing-library/dom'
 import { outdent } from 'outdent'
 
-import { initCheckboxes } from './checkboxes.mjs'
+import { Checkboxes, initCheckboxes } from './checkboxes.mjs'
 
 describe('Checkboxes', () => {
+  /** @type {HTMLElement} */
+  let $root
+
   /** @type {HTMLDivElement[]} */
   let $conditionals
 
@@ -89,21 +92,21 @@ describe('Checkboxes', () => {
       </form>
     `
 
-    const $container = document.querySelector('.nhsuk-checkboxes')
+    $root = document.querySelector('.nhsuk-checkboxes')
 
     $conditionals = [
-      ...$container.querySelectorAll('.nhsuk-checkboxes__conditional')
+      ...$root.querySelectorAll('.nhsuk-checkboxes__conditional')
     ]
 
-    const $input1 = getByRole($container, 'checkbox', {
+    const $input1 = getByRole($root, 'checkbox', {
       name: 'Email'
     })
 
-    const $input2 = getByRole($container, 'checkbox', {
+    const $input2 = getByRole($root, 'checkbox', {
       name: 'Phone'
     })
 
-    const $input3 = getByRole($container, 'checkbox', {
+    const $input3 = getByRole($root, 'checkbox', {
       name: 'Text message'
     })
 
@@ -114,7 +117,7 @@ describe('Checkboxes', () => {
     jest.spyOn($input3, 'addEventListener')
   })
 
-  describe('Initialisation', () => {
+  describe('Initialisation via init function', () => {
     it('should add event listeners', () => {
       initCheckboxes()
 
@@ -142,6 +145,34 @@ describe('Checkboxes', () => {
     it('should not throw with empty scope', () => {
       const scope = document.createElement('div')
       expect(() => initCheckboxes({ scope })).not.toThrow()
+    })
+  })
+
+  describe('Initialisation via class', () => {
+    it('should not throw with $root element', () => {
+      expect(() => new Checkboxes($root)).not.toThrow()
+    })
+
+    it('should throw with unsupported browser', () => {
+      document.body.classList.remove('nhsuk-frontend-supported')
+
+      expect(() => new Checkboxes($root)).toThrow(
+        'NHS.UK frontend is not supported in this browser'
+      )
+    })
+
+    it('should throw with missing $root element', () => {
+      expect(() => new Checkboxes()).toThrow(
+        'Checkboxes: Root element (`$root`) not found'
+      )
+    })
+
+    it('should throw with wrong $root element type', () => {
+      $root = document.createElementNS('http://www.w3.org/2000/svg', 'svg')
+
+      expect(() => new Checkboxes($root)).toThrow(
+        'Checkboxes: Root element (`$root`) is not of type HTMLElement'
+      )
     })
   })